Improve remove_untranslated.sh

+++ scripts/remove_untranslated.sh
@@ -5,6 +5,8 @@
 
 set -ex
 
+trap "rm -vf $po" 1 2 3 6
+
 if [ -n "$1" ]; then
   cd "$1"
 fi 
@@ -12,9 +14,14 @@ fi
 pofiles=$(find * -name '*.po' | sort)
 to_remove=()
 for po in $pofiles; do
-  output=$(LC_ALL=C /usr/bin/msgfmt -cvo /dev/null $po 2>&1 | grep '[0-9] translated messages')
+  output=$(LC_ALL=C /usr/bin/msgfmt -cvo /dev/null $po 2>&1 | grep -E '[0-9] translated message')
   if $(echo $output | grep '^0 translated messages' > /dev/null); then
     to_remove+=($po)
   fi
 done
-git rm ${to_remove[@]}
+
+if [ -n "${to_remove[@]}" ]; then
+  git rm ${to_remove[@]}
+else
+  echo "No empty PO to remove."
+fi
